bcm2835-bootloader: install dtbs only if avaible

+++ packages/tools/bcm2835-bootloader/scripts/update.sh
@@ -30,8 +30,14 @@
   cp -p $SYSTEM_ROOT/usr/share/bootloader/fixup.dat $BOOT_ROOT
   cp -p $SYSTEM_ROOT/usr/share/bootloader/start.elf $BOOT_ROOT
 
-  cp -p $SYSTEM_ROOT/usr/share/bootloader/*.dtb $BOOT_ROOT
-  cp -pR $SYSTEM_ROOT/usr/share/bootloader/overlays $BOOT_ROOT
+  if [ -f $SYSTEM_ROOT/usr/share/bootloader/*.dtb ]; then
+    cp -p $SYSTEM_ROOT/usr/share/bootloader/*.dtb $BOOT_ROOT
+  fi
+
+  if [ -f $SYSTEM_ROOT/usr/share/bootloader/overlays/*.dtb ]; then
+    mkdir -p $BOOT_ROOT/overlays
+    cp -p $SYSTEM_ROOT/usr/share/bootloader/overlays/*.dtb $BOOT_ROOT/overlays
+  fi
 
 # cleanup not more needed files
   rm -rf $BOOT_ROOT/loader.bin
